atlassian-tricked-out/requirements.txt
atlassian-python-api>=3.41.16
beautifulsoup4>=4.12.3
httpx>=0.28.0
mcp>=1.0.0
python-dotenv>=1.0.1
markdownify>=0.11.6
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
nltk>=3.8.0
joblib>=1.3.0
pyjwt>=2.7.0
requests>=2.31.0